Change v2 volunteer endpoint from /v2/volunteer to /v2/volunteers
