+++ controllers/v2/__init__.py
@@ -1,2 +1,2 @@
-from .volunteer import *
+from .volunteers import *
 from .v2_blueprint import v2_api, v2_bp
